Can successfully parse Tomb Raider 2 files.
Writing isn't tested yet; personally, I have more doubt here, but not that much. The unit test requires the app store version of Tomb Raider 2 to be installed and reads the first level (The Great Wall), which should be a good example of a level including a little bit of everything that could be annoying.