Fixed writing of levels (and added test).
The poser will not write level files, but I like to have the option for the future. I've now added a check whether I can write levels and then read them again (not yet checking whether the TR engine will accept them; that's a whole different thing), and this allowed to fix a few bugs. In particular:
- The geometry sub area in a room only goes up to Sprites, but does not include portals or anything after them.
- The TR 2 room vertex did not call the super implementation of writeToStream, even though I had clearly intended that and just forgot.