Transform meshes to SCNGeometry
SceneKit is interesting, but I'm not yet sure if it really is what I want. But it seems close enough, and if I decide to change to something written myself later, I can always reuse most of the work here.

The preview window is kind of psychedelic; that seems to be a decision by SceneKit (I don't do anything there). Also, everything is black, which I find very confusing.