Take the xCoord/yCoord property into account to reduce seams.
Texture coordinates in TR lie exactly on the boundaries of the textures. To avoid bleeding through of other textures, the x/y coord attributes are 1 when at the lower edge and 255 when at the upper, to allow for adjustments. The simplest adjustment is shifting the pixel by 1, which is probably less complex than what Core did (otherwise, why did they make the pixels lie on the boundaries in the first place?), but works very well.