Adjusted lighting of static meshes.
Static meshes that have a color set now get this in SceneKit as well. To do this, the geometry and the materials have to be copied. A more intelligent implementation might set up dictionaries for both by light value, to avoid extraneous copies, but so far, I don't see the need for this. I'm also less and less sure that SceneKit is really the solution for me here.