Added rendering of full levels.
That seemed easy! But it isn't, because it will horribly fail when rooms overlap each other (such as the TR3 mansion, everywhere). This is not a tiny issue here and there; it happens all over the place, especially in Tomb Raider 3, and is completely unacceptable. The correct solution is to implement portal rendering, something I'm afraid of at the best of times, and doing it with SceneKit is going to be very, very hard.

I'm thinking about doing some sort of fork; with scene kit rendering for basic stuff, and hand-written, suboptimal OpenGL for really dealing with levels.