Save the level, and change all lights.
Saving the level is a big one, because I didn't test my saving code in the engine before, but it works admirably - not a big surprise given that the entire system was specifically built for that, but nice. The important thing here was to set all room vertices to the same light value, because the ones used when following TRosettaStone just don't make any sense and I wanted to experiment.