New light scaling code.
The TRosetta stone lists the TR2 color scale as ranging from 0 (maximum brightness) to 8192 (minimum brightness). While technically true, this mapping is not linear; using it linearly will generate levels that are too dark. Through experimentation, I reached the conclusion that the real range is 4096-8192, with anything < 4096 having the same lighting as 4096. Thus this formula. I think it may still generate incorrect values, but at least not as incorrect as they were.