Deleted class because light normalization happens in the level.
I first thought that light normalization should take place in the mesh.
But that won't work; light values are used all over the place.
Normalizing it is thus the work of the level, which means that TR3Mesh
doesn't need this method anymore, and without any method, it's lost its
right to exist.