Checking TR4 start.
The TR4 start is indicated by 'T', 'R', '4', '\0'. I could make a 32-bit number out of that, as Apple used to back in the OS9 days, but writing it one by one and as the actual characters feels much nicer. That required some new code in the scanner additions, to deal with integers specified that way.