Corrected count, do not calculate width twice.
The capacity of an array is not that important, but I prefer to set it correctly or not at all. Tomb Raider 4 has no 8 bit texture tiles, so now the bit-agnostic accessor is used here.
Also, I was calculating the page width and height here again, for no apparent reason (actually it's a copy-and-paste leftover). That was stupid.