Ensure that offsets are converted correctly.
The most important part here is the .0 at the last 1024. Without it, this was an integer division that was then converted to float, losing precision in the process. I've now added casts to all of these things, to ensure that this doesn't happen again and that I know for a fact that I'm doing a conversion here when looking at it.