Reorganized the source code for inclusion in other projects.
I have for a long time planned to include parts of this in other, future software. Which may actually be finished before this program, but that's life for you. This recent change makes it a bit easier by hiding all dependencies on SceneKit for those who don't want to depend on it.